74720: No more submission-forms-migrate script access via REST
